(#6) Updated signature of showHighlightWindow

+++ src/highlightwindow.h
@@ -2,6 +2,8 @@
 #ifndef HIGHLIGHT_WINDOW_H
 #define HIGHLIGHT_WINDOW_H
 
-void showHighlightWindow(int x, int y, int width, int height, int duration, float opacity);
+#include <stdint.h>
+
+void showHighlightWindow(int32_t x, int32_t y, int32_t width, int32_t height, long duration, float opacity);
 
 #endif
